Moved SchemaValidator to unit tests
