+++ caom2/caom2/obs_reader_writer.py
@@ -77,7 +77,6 @@ import os
 import uuid
 from builtins import str, int
 import six
-from six import BytesIO
 from six.moves.urllib.parse import urlparse
 
 from lxml import etree
@@ -2510,54 +2509,3 @@ class ObservationWriter(object):
 
 class ObservationParsingException(Exception):
     pass
-
-
-class SchemaValidator(object):
-    """ This class provides a means to validate the XML schema. """
-
-    def _create_observation(self):
-        obs = observation.SimpleObservation(
-            "SCHEMA_VALIDATOR_COLLECTION", "schemaValidatorObsID")
-        obs.intent = observation.ObservationIntentType.SCIENCE
-        return obs
-
-    def _write_observation(self, observation):
-        writer = ObservationWriter(True, False, "caom2", CAOM23_NAMESPACE)
-        output = BytesIO()
-        writer.write(observation, output)
-        xml = output.getvalue()
-        output.close()
-        return xml
-
-    def _validate_with_intent_type(self, intent_type=None):
-        valid_observation = self._create_observation()
-        xml = self._write_observation(valid_observation)
-        if intent_type is not None:
-            xml = xml.replace(b'science', intent_type)
-        reader = ObservationReader(True)
-        reader.read(BytesIO(xml))
-
-    @staticmethod
-    def validate_schema():
-        """
-        This function is to catch unsupported extensions to our schema.
-        This function validates the schema by performing two simple observation
-        read-write round trips. One round trip uses a valid observation while
-        another uses an observation with an invalid value. The former should
-        pass while the latter should fail.
-        """
-        myself = SchemaValidator()
-        try:
-            # use a valid observation, should not catch any error
-            myself._validate_with_intent_type()
-        except Exception as ex:
-            raise AssertionError('Schema error: {}'.format(str(ex)))
-
-        try:
-            # use an invalid observation, should catch any error
-            myself._validate_with_intent_type(b'nosuchintent')
-            raise AssertionError(
-                'Schema failed to detect an error in ObservationIntentType.')
-        except Exception as ex:
-            if 'nosuchintent' not in str(ex):
-                raise AssertionError('Schema error: {}'.format(str(ex)))
